# Frontend/client/openapi.json
{"openapi":"3.1.0","info":{"title":"GitHub Analysis Platform","version":"0.1.0"},"paths":{"/api/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_api_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignup"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/change-password":{"post":{"tags":["Authentication"],"summary":"Change Password","operationId":"change_password_api_auth_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/me":{"get":{"tags":["User"],"summary":"Get Current User Profile","operationId":"get_current_user_profile_api_user_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/keys":{"get":{"tags":["User"],"summary":"Get Keys","operationId":"get_keys_api_user_keys_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["User"],"summary":"Add Key","operationId":"add_key_api_user_keys_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LLMKeyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/keys/{key_id}":{"delete":{"tags":["User"],"summary":"Delete Key","operationId":"delete_key_api_user_keys__key_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"key_id","in":"path","required":true,"schema":{"type":"string","title":"Key Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user/llm-key":{"get":{"tags":["User"],"summary":"Get Llm Key","operationId":"get_llm_key_api_user_llm_key_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["User"],"summary":"Save Llm Key","operationId":"save_llm_key_api_user_llm_key_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LLMKeyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/user/repos":{"get":{"tags":["User"],"summary":"Get User Repos","operationId":"get_user_repos_api_user_repos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/repo/analyze":{"post":{"tags":["Repository"],"summary":"Analyze Repo","operationId":"analyze_repo_api_repo_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_analyze_repo_api_repo_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/repo/{repo_id}/summary":{"get":{"tags":["Repository"],"summary":"Get Repo Summary","operationId":"get_repo_summary_api_repo__repo_id__summary_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/files":{"get":{"tags":["Repository"],"summary":"Get Repo Files","operationId":"get_repo_files_api_repo__repo_id__files_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/user/all":{"get":{"tags":["Repository"],"summary":"Get User Repos","operationId":"get_user_repos_api_repo_user_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/repo/{repo_id}/contributors":{"get":{"tags":["Repository"],"summary":"Get Contributors","operationId":"get_contributors_api_repo__repo_id__contributors_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/file/analyze":{"post":{"tags":["Repository"],"summary":"Analyze File","operationId":"analyze_file_api_repo__repo_id__file_analyze_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileAnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/file/content":{"post":{"tags":["Repository"],"summary":"Get File Content","operationId":"get_file_content_api_repo__repo_id__file_content_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileAnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/commits":{"get":{"tags":["Commits"],"summary":"Get Commits","operationId":"get_commits_api_repo__repo_id__commits_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/commits/story":{"get":{"tags":["Commits"],"summary":"Get Commit Stories","operationId":"get_commit_stories_api_repo__repo_id__commits_story_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/repo/{repo_id}/story":{"get":{"tags":["Commits"],"summary":"Get Project Story","operationId":"get_project_story_api_repo__repo_id__story_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"repo_id","in":"path","required":true,"schema":{"type":"string","title":"Repo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/llm/models":{"get":{"tags":["LLM"],"summary":"Get Available Models","operationId":"get_available_models_api_llm_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_analyze_repo_api_repo_analyze_post":{"properties":{"repo_url":{"type":"string","title":"Repo Url"}},"type":"object","required":["repo_url"],"title":"Body_analyze_repo_api_repo_analyze_post"},"ChangePasswordRequest":{"properties":{"old_password":{"type":"string","title":"Old Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["old_password","new_password"],"title":"ChangePasswordRequest"},"FileAnalysisRequest":{"properties":{"file_path":{"type":"string","title":"File Path"}},"type":"object","required":["file_path"],"title":"FileAnalysisRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LLMKeyRequest":{"properties":{"name":{"type":"string","title":"Name","default":"My Key"},"provider":{"type":"string","title":"Provider","default":"gemini-2.5-flash"},"api_key":{"type":"string","title":"Api Key"}},"type":"object","required":["api_key"],"title":"LLMKeyRequest"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserSignup":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserSignup"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}